app tsx 전환
